requirements.txt
streamlit
python-dotenv
google-generativeai
streamlit-mermaid
